refactor: separate library to `pipx-install-action` workspace
