+++ .gitignore
@@ -4,7 +4,6 @@
 !.git*
 
 coverage/
-dist/lib/
 node_modules/
 
 package.tgz
