+++ package.json
@@ -1,43 +1,19 @@
 {
-  "name": "pipx-install-action",
-  "version": "0.1.0",
-  "description": "Install Python packages using pipx with cache support on GitHub Actions",
-  "keywords": [
-    "action",
-    "python",
-    "install",
-    "cache",
-    "pipx"
-  ],
-  "homepage": "https://github.com/threeal/pipx-install-action#readme",
-  "bugs": {
-    "url": "https://github.com/threeal/pipx-install-action/issues",
-    "email": "alfi.maulana.f@gmail.com"
-  },
-  "repository": "github:threeal/pipx-install-action",
-  "license": "MIT",
-  "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
   "type": "module",
-  "main": "dist/lib/index.js",
-  "types": "dist/lib/index.d.ts",
-  "bin": "dist/lib/main.js",
-  "files": [
-    "dist/lib"
+  "workspaces": [
+    "pipx-install-action"
   ],
   "scripts": {
-    "build": "ncc build src/main.ts --out dist/action",
+    "build": "ncc build src/index.ts",
     "format": "prettier --write --cache . !dist",
     "lint": "eslint src",
-    "prepack": "tsc --declaration",
     "test": "jest"
   },
   "dependencies": {
-    "@actions/cache": "^3.2.4",
     "@actions/core": "^1.10.1",
-    "@actions/exec": "^1.1.1"
+    "pipx-install-action": "workspace:^"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.11.24",
     "@typescript-eslint/eslint-plugin": "^7.1.0",
