+++ pipx-install-action/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "pipx-install-action",
+  "version": "0.1.0",
+  "description": "Install Python packages using pipx with cache support on GitHub Actions",
+  "keywords": [
+    "action",
+    "python",
+    "install",
+    "cache",
+    "pipx"
+  ],
+  "homepage": "https://github.com/threeal/pipx-install-action#readme",
+  "bugs": {
+    "url": "https://github.com/threeal/pipx-install-action/issues",
+    "email": "alfi.maulana.f@gmail.com"
+  },
+  "repository": "github:threeal/pipx-install-action",
+  "license": "MIT",
+  "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "bin": "dist/main.js",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "prepack": "tsc"
+  },
+  "dependencies": {
+    "@actions/cache": "^3.2.4",
+    "@actions/core": "^1.10.1",
+    "@actions/exec": "^1.1.1"
+  },
+  "devDependencies": {
+    "@jest/globals": "^29.7.0",
+    "@types/node": "^20.11.24",
+    "typescript": "^5.3.3"
+  }
+}
