+++ pipx-install-action/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "include": ["src"],
+  "exclude": ["**/*.test.*"],
+  "compilerOptions": {
+    "exactOptionalPropertyTypes": true,
+    "strict": true,
+    "module": "node16",
+    "moduleResolution": "node16",
+    "declaration": true,
+    "outDir": "dist",
+    "esModuleInterop": true,
+    "target": "es2022",
+    "skipLibCheck": true
+  }
+}
