+++ tsconfig.json
@@ -1,14 +1,9 @@
 {
   "include": ["src"],
-  "exclude": ["**/*.test.*"],
   "compilerOptions": {
-    "exactOptionalPropertyTypes": true,
-    "strict": true,
     "module": "node16",
     "moduleResolution": "node16",
-    "outDir": "dist/lib",
     "esModuleInterop": true,
-    "target": "es2022",
-    "skipLibCheck": true
+    "target": "es2022"
   }
 }
